requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
autopep8=1.5.7=pypi_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py39h06a4308_0
charset-normalizer=2.0.4=pypi_0
click=8.0.1=pypi_0
et-xmlfile=1.1.0=pypi_0
flake8=3.9.2=pypi_0
idna=3.2=pypi_0
jdcal=1.4.1=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
mccabe=0.6.1=pypi_0
ncurses=6.2=he6710b0_1
numpy=1.21.1=pypi_0
openpyxl=2.4.11=pypi_0
openssl=1.1.1k=h27cfd23_0
pandas=1.3.1=pypi_0
pip=21.2.2=py39h06a4308_0
pycodestyle=2.7.0=pypi_0
pyflakes=2.3.1=pypi_0
python=3.9.6=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
selenium=3.141.0=pypi_0
setuptools=52.0.0=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
tzdata=2021a=h52ac0ba_0
urllib3=1.26.6=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xlsx2csv=0.7.8=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
